+++ src/components/tree/Download/index.tsx
@@ -2,10 +2,10 @@ import useUser from "../../../hooks/useUser";
 import { Analysis } from "../../../types/Analysis";
 import { TreeResult } from "../../../types/result";
 import FloatingButtonMenu from "../../common/FloatingButtonMenu";
+import FloatingButtonMenuItems from "../../common/FloatingButtonMenuItems";
 import AddIcon from "../../vectors/AddIcon";
 import CancelIcon from "../../vectors/CancelIcon";
 import DownloadIcon from "../../vectors/Download";
-import FloatingButtonMenuItems from "./FloatingButtonMenuItems";
 import { FC } from "react";
 import { Link } from "react-router-dom";
